Wayland: Fix event pump.
It was possible for the UntilEmpty event pump function to stay in the while loop, this also checks to see if the dispatch_pending functions errors out and properly sends out the error on the EBus for pretty printing if need be.

A simple fix for pointer to not apply cursor state if we don't have a valid serial.

Signed-off-by: Reece Hagan <drogonmar@tuyuji.com>